Fix ALOGVs for 64bit

+++ libvideoeditor/vss/stagefrightshells/src/VideoEditorMp3Reader.cpp
@@ -335,7 +335,7 @@ M4OSA_ERR VideoEditorMp3Reader_setOption(M4OSA_Context context,
     M4OSA_ERR err = M4NO_ERROR;
 
     ALOGV("VideoEditorMp3Reader_Context begin: optionId: %d Value: %d ",
-        (int)optionId,(int)pValue);
+        (int)optionId,(int)(intptr_t)pValue);
 
     M4OSA_DEBUG_IF1((M4OSA_NULL == pReaderContext), M4ERR_PARAMETER,
         "invalid context pointer");
